Fix bug in oracle update where wrong virtual offsets were used
This almost certainly would've led to wrong values being reported.

Problem was; `_getVirtualParameters()` with no arguments used the old `_lastInvariant`, not the current one. This is wrong and trickled down to the oracle update. The offending function is removed now. Re-organized the code to get the right virtual reserve offsets where they need to be without an unnecessary recompute of the invariant.